pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TiDHy"
version = "0.1.0"
description = "Time-Dependent Hybrid Dynamical Systems: Deep learning models for dynamical systems with discrete latent states using JAX/Flax"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Elliott Abe", email = "eabe@example.com"}
]
keywords = ["dynamical systems", "deep learning", "jax", "flax", "time series", "hybrid systems"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies
dependencies = [
    # Core scientific computing
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",

    # JAX ecosystem
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "flax>=0.8.0",
    "optax>=0.1.7",
    "orbax-checkpoint>=0.4.0",
    "chex>=0.1.8",

    # ML/Dynamical Systems
    "dynamax>=1.0.0",
    "scikit-learn>=1.3.0",
    "tensorflow-probability>=0.24.0",  # Requires tfp_jax_patch for JAX 0.8+

    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",

    # Experiment tracking
    "wandb",

    # Data I/O
    "h5py>=3.8.0",
    "tables>=3.8.0",  # pytables

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Utilities
    "tqdm>=4.65.0",
    "natsort",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

# Notebook dependencies
notebooks = [
    "jupyter",
    "jupyterlab",
    "ipykernel",
    "ipywidgets",
]

# SSM baselines
ssm = [
    "ssm @ git+https://github.com/lindermanlab/ssm@main",
    # Note: tensorflow-probability has compatibility issues with JAX 0.8+
    # Install manually if needed: pip install tensorflow-probability==0.24.0 'jax[cuda12]<0.5'
    "autograd>=1.7.0",
]

# RAPIDS support (for GPU-accelerated dataframes/ML)
rapids = [
    # Note: Install RAPIDS via conda, not pip
    # This is a placeholder for documentation
]

[project.urls]
Homepage = "https://github.com/elliottabe/TiDHy"
Repository = "https://github.com/elliottabe/TiDHy"
Documentation = "https://github.com/elliottabe/TiDHy/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["TiDHy*"]
exclude = ["tests*", "docs*", "configs*"]

[tool.setuptools.package-data]
TiDHy = ["py.typed"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | _build
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

# mypy configuration
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=TiDHy --cov-report=term-missing"

# Pyright/Pylance configuration
[tool.pyright]
typeCheckingMode = "off"
reportCallIssue = "none"
reportUnnecessaryLambda = "none"
reportGeneralTypeIssues = "none"
reportReturnType = "none"
reportArgumentType = "none"
reportAssignmentType = "none"
reportAttributeAccessIssue = "none"
reportPossiblyUnboundVariable = "none"
